+++ mycodo/config.py
@@ -7,8 +7,8 @@ import collections
 from datetime import timedelta
 from flask_babel import lazy_gettext
 
-MYCODO_VERSION = '5.4.15'
-ALEMBIC_VERSION = '8b36095c6cf9'
+MYCODO_VERSION = '5.4.16'
+ALEMBIC_VERSION = 'd36de7e4e477'
 
 LANGUAGES = {
     'en': 'English',