+++ mycodo/mycodo_flask/utils/utils_conditional.py
@@ -14,6 +14,7 @@ from mycodo.databases.models import Camera
 from mycodo.databases.models import Conditional
 from mycodo.databases.models import ConditionalActions
 from mycodo.databases.models import Input
+from mycodo.databases.models import Math
 from mycodo.utils.system_pi import is_int
 
 from mycodo.mycodo_flask.utils.utils_general import delete_entry_with_id
@@ -22,12 +23,14 @@ from mycodo.mycodo_flask.utils.utils_general import flash_success_errors
 logger = logging.getLogger(__name__)
 
 
-def conditional_add(cond_type, quantity, sensor_id=None):
+def conditional_add(cond_type, quantity, sensor_id=None, math_id=None):
     error = []
     if cond_type == 'relay':
         conditional_type = gettext(u"Output")
     elif cond_type == 'sensor':
         conditional_type = gettext(u"Input")
+    elif cond_type == 'math':
+        conditional_type = gettext(u"Math")
     else:
         error.append("Unrecognized conditional type: {cond_type}".format(
             cond_type=cond_type))
@@ -43,8 +46,11 @@ def conditional_add(cond_type, quantity, sensor_id=None):
                 new_conditional = Conditional()
                 try:
                     new_conditional.conditional_type = cond_type
+                    # TODO: Change to unique_id in next major version
                     if sensor_id:
                         new_conditional.sensor_id = sensor_id
+                    if math_id:
+                        new_conditional.math_id = math_id
                     new_conditional.save()
                 except sqlalchemy.exc.OperationalError as except_msg:
                     error.append(except_msg)
@@ -53,7 +59,13 @@ def conditional_add(cond_type, quantity, sensor_id=None):
 
                 if cond_type == 'sensor':
                     check_refresh_conditional(
+                        'sensor',
                         sensor_id,
+                        'add')
+                elif cond_type == 'math':
+                    check_refresh_conditional(
+                        'math',
+                        math_id,
                         'add')
     flash_success_errors(error, action, url_for('page_routes.page_output'))
 
@@ -66,6 +78,8 @@ def conditional_mod(form, mod_type):
         cond_type = gettext(u"Output")
     elif conditional_type == 'sensor':
         cond_type = gettext(u"Input")
+    elif conditional_type == 'math':
+        cond_type = gettext(u"Math")
     else:
         error.append("Unrecognized conditional type: {cond_type}".format(
             cond_type=form.conditional_type.data))
@@ -87,6 +101,12 @@ def conditional_mod(form, mod_type):
 
             if conditional_type == 'sensor':
                 check_refresh_conditional(
+                    'sensor',
+                    form.sensor_id.data,
+                    'del')
+            elif conditional_type == 'math':
+                check_refresh_conditional(
+                    'math',
                     form.sensor_id.data,
                     'del')
 
@@ -110,6 +130,11 @@ def conditional_mod(form, mod_type):
                     mod_action.if_sensor_gpio_state = form.if_sensor_gpio_state.data
                     mod_action.if_sensor_direction = form.if_sensor_direction.data
                     mod_action.if_sensor_setpoint = form.if_sensor_setpoint.data
+                elif conditional_type == 'math':
+                    mod_action.if_sensor_period = form.if_sensor_period.data
+                    mod_action.if_sensor_measurement = form.if_sensor_measurement.data
+                    mod_action.if_sensor_direction = form.if_sensor_direction.data
+                    mod_action.if_sensor_setpoint = form.if_sensor_setpoint.data
                 db.session.commit()
             except sqlalchemy.exc.OperationalError as except_msg:
                 error.append(except_msg)
@@ -118,8 +143,15 @@ def conditional_mod(form, mod_type):
 
             if conditional_type == 'sensor':
                 check_refresh_conditional(
+                    'sensor',
                     form.sensor_id.data,
                     'mod')
+            elif conditional_type == 'math':
+                check_refresh_conditional(
+                    'math',
+                    form.sensor_id.data,
+                    'mod')
+
     flash_success_errors(error, action, url_for('page_routes.page_output'))
 
 
@@ -131,6 +163,8 @@ def conditional_action_add(form):
         cond_type = gettext(u"Output")
     elif conditional_type == 'sensor':
         cond_type = gettext(u"Input")
+    elif conditional_type == 'math':
+        cond_type = gettext(u"Math")
     else:
         error.append("Unrecognized conditional type: {cond_type}".format(
             cond_type=form.conditional_type.data))
@@ -160,6 +194,8 @@ def conditional_action_mod(form, mod_type):
         cond_type = gettext(u"Output")
     elif cond.conditional_type == 'sensor':
         cond_type = gettext(u"Input")
+    elif cond.conditional_type == 'math':
+        cond_type = gettext(u"Math")
     else:
         error.append("Unrecognized conditional type: {cond_type}".format(
             cond_type=form.conditional_type.data))
@@ -230,38 +266,65 @@ def conditional_action_mod(form, mod_type):
 
         if cond.conditional_type == 'sensor':
             check_refresh_conditional(
+                'sensor',
                 cond.sensor_id,
+                'mod')
+        elif cond.conditional_type == 'math':
+            check_refresh_conditional(
+                'math',
+                cond.math_id,
                 'mod')
     flash_success_errors(error, action, url_for('page_routes.page_output'))
 
 
 def conditional_activate(form):
-    conditional = Conditional.query.filter(
+    cond = Conditional.query.filter(
         Conditional.id == form.conditional_id.data).first()
-    conditional.is_activated = True
+    cond.is_activated = True
     db.session.commit()
-    if conditional.conditional_type == 'sensor':
+    if cond.conditional_type == 'sensor':
         check_refresh_conditional(
+            'sensor',
+            form.sensor_id.data,
+            'mod')
+    elif cond.conditional_type == 'math':
+        check_refresh_conditional(
+            'math',
             form.sensor_id.data,
             'mod')
 
 
 def conditional_deactivate(form):
-    conditional = Conditional.query.filter(
+    cond = Conditional.query.filter(
         Conditional.id == form.conditional_id.data).first()
-    conditional.is_activated = False
+    cond.is_activated = False
     db.session.commit()
-    if conditional.conditional_type == 'sensor':
+    if cond.conditional_type == 'sensor':
         check_refresh_conditional(
+            'sensor',
+            form.sensor_id.data,
+            'mod')
+    elif cond.conditional_type == 'math':
+        check_refresh_conditional(
+            'math',
             form.sensor_id.data,
             'mod')
 
 
-def check_refresh_conditional(sensor_id, cond_mod):
-    sensor = (Input.query
-              .filter(Input.id == sensor_id)
-              .filter(Input.is_activated == True)
-              ).first()
-    if sensor:
-        control = DaemonControl()
-        control.refresh_sensor_conditionals(sensor_id, cond_mod)
+def check_refresh_conditional(sensor_id, cont_type, cond_mod):
+    if cont_type == 'sensor':
+        sensor = (Input.query
+                  .filter(Input.id == sensor_id)
+                  .filter(Input.is_activated == True)
+                  ).first()
+        if sensor:
+            control = DaemonControl()
+            control.refresh_sensor_conditionals(sensor_id, cond_mod)
+    if cont_type == 'math':
+        math = (Math.query
+                .filter(Input.id == sensor_id)
+                .filter(Input.is_activated == True)
+                ).first()
+        if math:
+            control = DaemonControl()
+            control.refresh_math_conditionals(sensor_id, cond_mod)